Fixed several issues in realtime processing Fixed detection3D not preallocating tracker vector - Expecting one extra space at the end of synchronous processing - Asynchronous detection uses it to append without reallocation - Required allowing a probe AND a retroactive tracking record     - Essentially allowing for multiple entries into tracker records each frame     - Readers have been updated to properly handle probing records Fixed inconsistent frame selection in realtime processing - Intended to skip frames only when delayed - Accidentally always used latest, resulting in potentially excessive skipping - But pushed older, potentially skipped frame to IO protocols Fixed realtime processing waiting until next frame, even with frame already waiting - This resulted in forced frame skips Added better warning to SOF when camera is unexpectedly not synced - Relevant if camera was not set up properly (e.g. comms failed) - If it is not syncing to SyncGroup, it may create SOF frames ahead of controller - While not fixed, this adds a warning that this happens - Workaround is restarting streaming, solution is making UART reliable Fixed overflows in controller event parsing due to type mismatch

+++ server/source/device/tracking_controller.hpp
@@ -44,7 +44,7 @@ const char *getControllerEventName(ControllerEventID event);
 struct ControllerEventLog
 {
 	uint64_t timestamp;
-	double timestampUS;
+	uint64_t timestampUS;
 	TimePoint_t syncedTime;
 	ControllerEventID id;
 	bool isNewEvent;
